Improve code readability
